Added Moveable class. Restructured file structure. Going to branch to implement.
