+++ classes/GameObject/Moveable/Moveable.h
@@ -1,0 +1,12 @@
+#ifndef MOVEABLE_H_
+#define MOVEABLE_H_
+
+#include "../GameObject.h"
+
+class Moveable : public GameObject {
+   public:
+
+   private:
+};
+
+#endif /* MOVEABLE_H_ */
